add lock for Actions on bare Pod
