+++ pkg/action/util/lock_helper.go
@@ -9,19 +9,21 @@ import (
 
 // a lock for bare pods to avoid concurrent contention of actions on the same pod.
 // detail of its purpose can be found: https://github.com/turbonomic/kubeturbo/issues/104
-type lockHelper struct {
+type LockHelper struct {
 	//for the expirationMap
 	emap    *ExpirationMap
 	key     string
 	version int64
+
+	callback expireCallBack
 
 	//stop Renewing
 	stop       chan struct{}
 	isRenewing bool
 }
 
-func NewLockHelper(podkey string, emap *ExpirationMap) (*lockHelper, error) {
-	p := &lockHelper{
+func NewLockHelper(podkey string, emap *ExpirationMap) (*LockHelper, error) {
+	p := &LockHelper{
 		key:        podkey,
 		emap:       emap,
 		stop:       make(chan struct{}),
@@ -37,30 +39,34 @@ func NewLockHelper(podkey string, emap *ExpirationMap) (*lockHelper, error) {
 	return p, nil
 }
 
-func (h *lockHelper) Setkey(key string) {
+func (h *LockHelper) Setkey(key string) {
 	h.key = key
 	return
 }
 
-func (h *lockHelper) Acquirelock() bool {
-	version, flag := h.emap.Add(h.key, nil, func(obj interface{}) {
-		h.lockCallBack()
-	})
+// the input of this callback will be h.key.
+func (h *LockHelper) AcquireLock(callback expireCallBack) bool {
+	if callback == nil {
+		callback = func(obj interface{}) {
+			h.lockCallBack()
+		}
+	}
+	version, flag := h.emap.Add(h.key, h.key, callback)
 
 	if !flag {
-		glog.V(3).Infof("Failed to get lock for pod [%s]", h.key)
+		glog.V(3).Infof("Failed to get lock for [%s]", h.key)
 		return false
 	}
 
-	glog.V(4).Infof("Get lock for pod [%s]", h.key)
+	glog.V(4).Infof("Get lock for [%s]", h.key)
 	h.version = version
 
 	return true
 }
 
-func (h *lockHelper) Trylock(timeout, interval time.Duration) error {
+func (h *LockHelper) Trylock(timeout, interval time.Duration) error {
 	err := goutil.RetryDuring(1000, timeout, interval, func() error {
-		if !h.Acquirelock() {
+		if !h.AcquireLock(nil) {
 			return fmt.Errorf("TryLater")
 		}
 		return nil
@@ -73,22 +79,22 @@ func (h *lockHelper) Trylock(timeout, interval time.Duration) error {
 	return nil
 }
 
-func (h *lockHelper) ReleaseLock() {
+func (h *LockHelper) ReleaseLock() {
 	h.emap.Del(h.key, h.version)
 	h.StopRenew()
-	glog.V(4).Infof("Released lock for pod [%s]", h.key)
+	glog.V(4).Infof("Released lock for [%s]", h.key)
 }
 
-func (h *lockHelper) lockCallBack() {
+func (h *LockHelper) lockCallBack() {
 	// do nothing
 	return
 }
 
-func (h *lockHelper) Renewlock() bool {
+func (h *LockHelper) RenewLock() bool {
 	return h.emap.Touch(h.key, h.version)
 }
 
-func (h *lockHelper) KeepRenewLock() {
+func (h *LockHelper) KeepRenewLock() {
 	ttl := h.emap.GetTTL()
 	interval := ttl / 2
 	if interval < time.Second {
@@ -103,7 +109,7 @@ func (h *lockHelper) KeepRenewLock() {
 				glog.V(3).Infof("schedulerHelper stop renewlock.")
 				return
 			default:
-				if !h.Renewlock() {
+				if !h.RenewLock() {
 					return
 				}
 				time.Sleep(interval)
@@ -113,7 +119,7 @@ func (h *lockHelper) KeepRenewLock() {
 	}()
 }
 
-func (h *lockHelper) StopRenew() {
+func (h *LockHelper) StopRenew() {
 	if h.isRenewing {
 		h.isRenewing = false
 		close(h.stop)
